Re Re Re Modificado clases querybuilder y query
